==> quest-generator/src/00-crawl_quests.ipynb <==
import requests
from bs4 import BeautifulSoup

quest_article = requests.get("https://elderscrolls.fandom.com/wiki/Falion%27s_Secret").text

soup = BeautifulSoup(quest_article, 'html.parser')

title = soup.find_all("h1")[0].getText()
title

soup.find_all("ol")

background_text = soup.find_all("div", {"class": "blockquote-half"})[0].getText()
background_text

objectives = soup.find_all("ol")[0].getText()
objectives
