==> quest-generator/src/00-get_skyrim_quests_xml.ipynb <==
import xml.etree.ElementTree as ET

tree = ET.parse('elderscrolls_pages_current.xml')
root = tree.getroot()

for i, e in enumerate(root.iter()):
    print(e.tag)
    if i == 100:
        break

skyrim_quests = root.find('./page/[title="Quests (Skyrim)"]')

quest_index_text = skyrim_quests.find('revision').find('text').text

print(quest_index_text)
