==> quest-generator/src/01-get_quests_api.ipynb <==
import requests
import json
import time

---

# Skyrim

skyrim_quests = json.loads(requests.get("https://elderscrolls.fandom.com/api/v1/Articles/List?category=Skyrim%3A_Quests&limit=500").text)

quest_ids = [item["id"] for item in skyrim_quests['items']]

all_skyrim_quests = {}

for i, quest_id in enumerate(quest_ids[1:]):
    try:
        print(f"{i}/{len(quest_ids)-1}")
        
        quest_article = json.loads(requests.get(f"https://elderscrolls.fandom.com/api/v1/Articles/AsSimpleJson?id={quest_id}").text)
        all_skyrim_quests[quest_id] = quest_article["sections"]
        
        time.sleep(5)
    except Exception as e:
        print(f"Failed to download quest {quest_id}: {e}")

with open("skyrim_quests.json", "w") as outfile:
    json.dump(all_skyrim_quests, outfile)

---

# The Witcher

witch_quests = json.loads(requests.get("https://witcher.fandom.com/api/v1/Articles/List?category=The_Witcher_3_main_quests&limit=500").text)

quest_ids = [item["id"] for item in witch_quests['items']]

all_witcher_quests = {}

for i, quest_id in enumerate(quest_ids[1:]):
    try:
        print(f"{i}/{len(quest_ids)-1}")
        
        quest_article = json.loads(requests.get(f"https://witcher.fandom.com/api/v1/Articles/AsSimpleJson?id={quest_id}").text)
        all_witcher_quests[quest_id] = quest_article["sections"]
        
        time.sleep(5)
    except Exception as e:
        print(f"Failed to download quest {quest_id}: {e}")

with open("witcher_quests.json", "w") as outfile:
    json.dump(all_witcher_quests, outfile)

## Contracts

witch_quests = json.loads(requests.get("https://witcher.fandom.com/api/v1/Articles/List?category=The_Witcher_3_contracts&limit=500").text)

quest_ids = [item["id"] for item in witch_quests['items']]

len(quest_ids)

all_witcher_contracts = {}

for i, quest_id in enumerate(quest_ids[1:]):
    try:
        print(f"{i}/{len(quest_ids)-1}")
        
        quest_article = json.loads(requests.get(f"https://witcher.fandom.com/api/v1/Articles/AsSimpleJson?id={quest_id}").text)
        all_witcher_contracts[quest_id] = quest_article["sections"]
        
        time.sleep(5)
    except Exception as e:
        print(f"Failed to download quest {quest_id}: {e}")

with open("witcher_contracts.json", "w") as outfile:
    json.dump(all_witcher_contracts, outfile)

## Side quests

witch_quests = json.loads(requests.get("https://witcher.fandom.com/api/v1/Articles/List?category=The_Witcher_3_secondary_quests&limit=500").text)

quest_ids = [item["id"] for item in witch_quests['items']]

len(quest_ids)

all_witcher_side_quests = {}

for i, quest_id in enumerate(quest_ids[1:]):
    try:
        print(f"{i}/{len(quest_ids)-1}")
        
        quest_article = json.loads(requests.get(f"https://witcher.fandom.com/api/v1/Articles/AsSimpleJson?id={quest_id}").text)
        all_witcher_side_quests[quest_id] = quest_article["sections"]
        
        time.sleep(5)
    except Exception as e:
        print(f"Failed to download quest {quest_id}: {e}")

with open("witcher_side_quests.json", "w") as outfile:
    json.dump(all_witcher_side_quests, outfile)

## Treasures

witch_quests = json.loads(requests.get("https://witcher.fandom.com/api/v1/Articles/List?category=The_Witcher_3_treasure_hunts&limit=500").text)

quest_ids = [item["id"] for item in witch_quests['items']]

len(quest_ids)

all_witcher_hunts = {}

for i, quest_id in enumerate(quest_ids[1:]):
    try:
        print(f"{i}/{len(quest_ids)-1}")
        
        quest_article = json.loads(requests.get(f"https://witcher.fandom.com/api/v1/Articles/AsSimpleJson?id={quest_id}").text)
        all_witcher_hunts[quest_id] = quest_article["sections"]
        
        time.sleep(5)
    except Exception as e:
        print(f"Failed to download quest {quest_id}: {e}")

with open("witcher_hunts.json", "w") as outfile:
    json.dump(all_witcher_hunts, outfile)
