==> quest-generator/src/02-generate_objectives_dataset.ipynb <==
import json
import pandas as pd

with open("quests/skyrim_quests.json") as infile:
    quests_json = json.load(infile)

quest_objectives = {
    "source": [],
    "id": [],
    "title": [],
    "objectives": []
}

for quest_id, quest in quests_json.items():
    try:
        quest_title = quest[0]["title"]
        quest_objs = ' \n'.join([e['text'] for e in quest[1]['content'][0]['elements']])

        quest_objectives["source"].append("skyrim")
        quest_objectives["id"].append(quest_id)    
        quest_objectives["title"].append(quest_title)    
        quest_objectives["objectives"].append(quest_objs)    
    except Exception as e:
        print(f"{quest_id} failed for {e}")

skyrim_quests = pd.DataFrame(quest_objectives)

---

with open("quests/witcher_contracts.json") as infile:
    quests_json = json.load(infile)

quest_objectives = {
    "source": [],
    "id": [],
    "title": [],
    "objectives": []
}

for quest_id, quest in quests_json.items():
    try:
        quest_title = quest[0]["title"]
        quest_objs = ' \n'.join([e['text'] for e in quest[2]['content'][0]['elements']])

        quest_objectives["source"].append("wicher_contracts")
        quest_objectives["id"].append(quest_id)    
        quest_objectives["title"].append(quest_title)    
        quest_objectives["objectives"].append(quest_objs)    
    except Exception as e:
        print(f"{quest_id} failed for {e}")

wicher_contracts = pd.DataFrame(quest_objectives)

wicher_contracts.head()

---

with open("quests/witcher_hunts.json") as infile:
    quests_json = json.load(infile)

quest_objectives = {
    "source": [],
    "id": [],
    "title": [],
    "objectives": []
}

for quest_id, quest in quests_json.items():
    try:
        quest_title = quest[0]["title"]
        quest_objs = ' \n'.join([e['text'] for e in quest[2]['content'][0]['elements']])

        quest_objectives["source"].append("wicher_contracts")
        quest_objectives["id"].append(quest_id)    
        quest_objectives["title"].append(quest_title)    
        quest_objectives["objectives"].append(quest_objs)    
    except Exception as e:
        print(f"{quest_id} failed for {e}")

wicher_hunts = pd.DataFrame(quest_objectives)

wicher_hunts.head()

---

with open("quests/witcher_quests.json") as infile:
    quests_json = json.load(infile)

quest_objectives = {
    "source": [],
    "id": [],
    "title": [],
    "objectives": []
}

for quest_id, quest in quests_json.items():
    try:
        quest_title = quest[0]["title"]
        for sec in quest:
            if sec["title"] == "Objectives":
                quest_objs = ' \n'.join([e['text'] for e in sec['content'][0]['elements']])

        quest_objectives["source"].append("wicher_contracts")
        quest_objectives["id"].append(quest_id)    
        quest_objectives["title"].append(quest_title)    
        quest_objectives["objectives"].append(quest_objs)    
    except Exception as e:
        print(f"{quest_id} failed for {e}")

wicher_quests = pd.DataFrame(quest_objectives)

wicher_quests.head()

---

with open("quests/witcher_side_quests.json") as infile:
    quests_json = json.load(infile)

quest_objectives = {
    "source": [],
    "id": [],
    "title": [],
    "objectives": []
}

for quest_id, quest in quests_json.items():
    try:
        quest_title = quest[0]["title"]
        for sec in quest:
            if sec["title"] == "Objectives":
                quest_objs = ' \n'.join([e['text'] for e in sec['content'][0]['elements']])

        quest_objectives["source"].append("wicher_contracts")
        quest_objectives["id"].append(quest_id)    
        quest_objectives["title"].append(quest_title)    
        quest_objectives["objectives"].append(quest_objs)    
    except Exception as e:
        print(f"{quest_id} failed for {e}")

wicher_side_quests = pd.DataFrame(quest_objectives)

wicher_side_quests.head()

---

quest_objectives_df = pd.concat([skyrim_quests, wicher_contracts, wicher_hunts, wicher_quests, wicher_side_quests])

quest_objectives_df.sample(5)

quest_objectives_df.shape

quest_objectives_df.to_csv("quests/quest_objectives.csv", sep=";")
