==> quest-generator/src/03a-extract_entities.ipynb <==
Ideia:
- Extract entities from the quests
- Replace entities with masks
- Train causal language model to generate quests if masks
- Use masked language modeling to fill masks in quests

https://huggingface.co/transformers/usage.html#named-entity-recognition

import json

import pandas as pd
from transformers import AutoModelForTokenClassification, AutoTokenizer, pipeline
import torch

from english_words import english_words_lower_set
from nltk.tokenize import word_tokenize

df = pd.read_csv("../data/quests/quest_objectives.csv", sep=";").drop(columns="Unnamed: 0")
df.head()

nlp = pipeline('ner', grouped_entities=True)

sentence = df.iloc[1]["objectives"]
print(sentence)

nlp(sentence)

entity_map = {
    "I-PER": "xperx",
    "I-ORG": "xorgx",
    "I-LOC": "xlocx",
    "I-MISC": "xmiscx"
}

token_entity_map = {}

for sentence in df["objectives"].values:
    for ent in nlp(sentence):
        if len(ent["word"]) > 1:
            token_entity_map[ent["word"]] = entity_map[ent["entity_group"]]

with open("../data/assets/entities.json", "w") as outfile:
    json.dump(token_entity_map, outfile)

---

def replace_sentence(sentence, token_map):
    return ' '.join([
        word if word not in token_map else token_map[word] 
        for word
        in word_tokenize(sentence)
    ])

df["objectives_entities"] = df["objectives"].apply(lambda x: replace_sentence(x, token_entity_map))

df.sample()["objectives_entities"].values

df.to_csv("../data/quests/quest_objectives_entities.csv", sep=";", index=False)

---
