==> quest-generator/src/03b-umlfit.ipynb <==
https://docs.fast.ai/tutorial.text#Fine-tuning-a-language-model-on-IMDb

import pickle

import pandas as pd
from sklearn.model_selection import train_test_split

from english_words import english_words_lower_set
import fastai
from fastai.text.all import *

df_quests = pd.read_csv("../data/quests/quest_objectives_entities.csv", sep=";")

df_quests.head()

df_quests["text"] = df_quests.apply(lambda x: (x["title"] + " > " + x["objectives_entities"].replace(".", " > ")).lower(), axis=1)

df_quests.sample(3)["text"]

with open("../data/raw/gameofthrones.txt") as infile:
    got_txt = infile.readlines()
got_txt = [l for l in got_txt if l != "\n"]

df_got = pd.DataFrame({"text": got_txt})

df = pd.concat([df_got["text"], df_quests["text"]]).to_frame()

lm_data = TextDataLoaders.from_df(df, text_col="text", is_lm=True, seed=7)

lm_data.show_batch(max_n=3)

learn = language_model_learner(lm_data, AWD_LSTM, metrics=[accuracy, Perplexity()], path="../models/")

learn.fit_one_cycle(1, 1e-2)

learn.save('1epoch')

learn = learn.load('1epoch')

learn.unfreeze()
learn.fit_one_cycle(10, 1e-3)

learn.export('finetuned_book.pkl')

with open("../models/models/book_vocab.pkl", "wb") as outfile:
    pickle.dump(lm_data.vocab, outfile)

---

df_quests.head()["title"]

quest_title = "Swords in the dark"

n_goals = 5
[learn.predict(quest_title.lower() + " >", n_words=15) for _ in range(n_goals)]

quest_title = "Swords in the dark > enter shrine"

n_goals = 5
[learn_quests.predict(quest_title.lower() + " >", n_words=15) for _ in range(n_goals)]

---

lm_data_quests = TextDataLoaders.from_df(df_quests, text_col="text", is_lm=True, seed=7, text_vocab=lm_data.vocab)

lm_data_quests.show_batch(max_n=3)

learn_quests = language_model_learner(lm_data_quests, AWD_LSTM, metrics=[accuracy, Perplexity()], path="../models/")

learn_quests.fit_one_cycle(1, 1e-2)

learn_quests.unfreeze()
learn_quests.fit_one_cycle(10, 1e-3)

learn_quests.export('finetuned_quests.pkl')

quest_title = "Swords in the dark"

n_goals = 5
[learn_quests.predict(quest_title.lower() + " >", n_words=15) for _ in range(n_goals)]

quest_title = "Swords in the dark > enter shrine"

n_goals = 5
[learn_quests.predict(quest_title.lower() + " >", n_words=15) for _ in range(n_goals)]

quest_title = "Swords in the dark > enter shrine > hunt there"

n_goals = 5
[learn_quests.predict(quest_title.lower() + " >", n_words=15) for _ in range(n_goals)]

---

# backup

learn.fine_tune(10, 1e-2)

learn.predict(title.lower() + ' > ', n_words=20)

learn.predict(title.lower() + ' > ', n_words=20)

title = "The Crimson Conspiracy"

learn.predict(title.lower() + ' > ', n_words=20)

title = "A curious looking egg"

learn.predict(title.lower() + ' > ', n_words=20)

---

loaded_learn = load_learner("../models/finetuned_quests")

loaded_learn
