==> quest-generator/src/03c-gpt2_lm.ipynb <==
import pandas as pd

from transformers import pipeline, set_seed

df = pd.read_csv("../data/quests/quest_objectives.csv", sep=";")

df.head()

generator = pipeline('text-generation', model='gpt2')
set_seed(42)

sample = df.sample() 

input_sentence = sample["title"].values[0] + " - " + "".join(sample["objectives"].values[0].split("\n")[0])
input_sentence

generator(input_sentence, max_length=50, num_return_sequences=5)
